Remove some classes in GrAITests.
