--- Tests/GrAITests/Base/MSE2DCase.swift
@@ -1,162 +1,0 @@
-//
-// MSE2DCase.swift
-// GrAITests
-//
-// Created by Jean-François Reboud on 05/03/2023.
-//
-
-import XCTest
-import GrAIdient
-import GrAITestsUtils
-
-///
-/// A class that will test a model with a structural hypothesis:
-/// the model last layer is a MSE2D layer.
-///
-class MSE2DCase: XCTestCase
-{
-    /// Batch size of data.
-    var batchSize: Int! = nil
-    /// Optimizer parameters.
-    var optimizerParams = GrAI.Optimizer.Params()
-    
-    /// Systematic call before test begins.
-    override func setUp()
-    {
-        batchSize = 5
-        _ = MetalKernel.get
-        GrAI.Opti.GPU = true
-        
-        setOptimizerParams(params: &optimizerParams)
-        optimizerParams.nbLoops = 3
-    }
-    
-    ///
-    /// Get the current batch size of data.
-    ///
-    /// This function allows to simulate the fact that the batch size of data may be smalling during the
-    /// last iteration of the training.
-    ///
-    /// - Parameter model: The model.
-    /// - Returns: The batch size of data.
-    ///
-    func getBatchSize(_ model: Model) -> Int
-    {
-        if model.optimizerParams.step == model.optimizerParams.nbLoops-1
-        {
-            return batchSize / 2
-        }
-        else
-        {
-            return batchSize
-        }
-    }
-    
-    ///
-    /// Create synthetic data.
-    ///
-    /// - Parameters:
-    ///     - dim1: The first dimension of the data.
-    ///     - dim2: The second dimension of the data.
-    /// - Returns: The created data.
-    ///
-    func buildData<T: BinaryFloatingPoint>(dim1: Int, dim2: Int) -> [[T]]
-    {
-        var data = [[T]]()
-        for _ in 0..<dim1
-        {
-            var data1 = [T]()
-            for _ in 0..<dim2
-            {
-                data1.append(T(Double.random(in: -1.0..<1.0)))
-            }
-            data.append(data1)
-        }
-        return data
-    }
-    
-    ///
-    /// A function to create/set ground truth to the model.
-    ///
-    /// - Parameters:
-    ///     - groundTruth: The ground truth to set.
-    ///     - model: The model.
-    /// - Returns: The ground truth.
-    ///
-    func setLoss(_ groundTruth: [[Double]]?, _ model: Model) -> [[Double]]
-    {
-        let lastLayer = model.layers.last as! MSE2D
-        let gt: [[Double]]
-        if let groundTruthTmp = groundTruth
-        {
-            gt = groundTruthTmp
-        }
-        else
-        {
-            gt = buildData(dim1: getBatchSize(model), dim2: 1)
-        }
-        
-        if GrAI.Opti.GPU
-        {
-            try! lastLayer.lossDerivativeGPU(
-                gt.reduce([], +), batchSize: gt.count, format: .Neuron
-            )
-        }
-        else
-        {
-            try! lastLayer.lossDerivativeCPU(
-                gt.reduce([], +), batchSize: gt.count, format: .Neuron
-            )
-        }
-        return gt
-    }
-    
-    ///
-    /// A function to get loss of a model.
-    ///
-    /// - Parameters:
-    ///     - groundTruth: The ground truth to set.
-    ///     - model: The model.
-    /// - Returns: The loss value.
-    ///
-    func getLoss(_ groundTruth: [[Double]], _ model: Model) -> Double
-    {
-        let lastLayer = model.layers.last as! MSE2D
-        if GrAI.Opti.GPU
-        {
-            return Double(try! lastLayer.getLossGPU(
-                groundTruth.reduce([], +),
-                batchSize: groundTruth.count,
-                format: .Neuron
-            ))
-        }
-        else
-        {
-            return try! lastLayer.getLossCPU(
-                groundTruth.reduce([], +),
-                batchSize: groundTruth.count,
-                format: .Neuron
-            )
-        }
-    }
-    
-    ///
-    /// A function to get the gradients of weights approximations..
-    ///
-    /// - Parameters:
-    ///     - groundTruth: The ground truth.
-    ///     - model: The model.
-    /// - Returns: The gradients of weights approximations.
-    ///
-    func getGradientsApprox(
-        _ groundTruth: [[Double]],
-        _ model: Model) -> [Double]
-    {
-        let lastLayer = model.layers.last as! MSE2D
-        return try! lastLayer.collectGradientsApprox(
-            groundTruth.reduce([], +),
-            batchSize: groundTruth.count,
-            format: .Neuron
-        )
-    }
-}
